Bring back heuristic fragment matching, with a twist.
Easily the biggest breaking change for fragment matching in Apollo Client
3.0 was the removal of the FragmentMatcher abstraction, along with its
subclasses HeuristicFragmentMatcher and IntrospectionFragmentmatcher,
which were replaced by the declarative possibleTypes configuration.

The HeuristicFragmentMatcher was so named because it attempted to perform
fragment matching without any actual knowledge of supertype/subtype
relationships in your schema, instead checking whether all the fields of a
given fragment were present in the result object, which is indeed a
relatively strong signal that the fragment probably matched.

The HeuristicFragmentMatcher could be fooled by aliases and accidental
sharing of field names between different fragments, but it was also
relatively resilient to the possibility of adding new subtypes on the
server, because heuristic matching doesn't care what the true subtypes of
a supertype are, so what's one more?

The other big drawback of the HeuristicFragmentMatcher was that it applied
the same fuzzy logic to reading from the cache, where the heuristic makes
a lot less sense. If an individual query result has all the keys you'd
expect if a certain fragment matched, that's a pretty good sign that the
fragment matched. But when you have a lot of data in your cache, from lots
of different queries, it's not as meaningful to observe that all the
fields required by the fragment are present, since they might be there
because they've been written into the cache by other queries over time,
not because the fragment actually matches the __typename of the object.

In moving to the more exact possibleTypes system, we gave up both the
benefits and the drawbacks of the HeuristicFragmentMatcher, replacing it
with something functionally similar to the IntrospectionFragmentmatcher,
but with a much simpler configuration API.

As #5750 demonstrates, the exactness of the possibleTypes system makes it
challenging for existing clients to adapt when a new subtype is added on
the server. Is there any way to make this system more flexible, like the
HeuristicFragmentMatcher, but without the drawbacks?

This PR improves the possibleTypes API by allowing "fuzzy" subtype
strings, which are interpreted as regular expressions for typenames,
rather than as actual typename strings. If a fuzzy subtype matches the
__typename of an object, fragments on supertypes of that fuzzy subtype are
allowed to match the object, provided the result also has all the keys
required by the fragment.

The key advantages of this new system compared to the
HeuristicFragmentMatcher are that (1) you can specify fuzzy subtypes for
specific supertypes (rather than applying heuristic matching for all types
by default), and (2) that it "learns" about fuzzy subtypes while writing
(where the heuristic tends to work very well), and then merely uses those
inferred supertype/subtype relationships when reading, so there is no need
to perform heuristic matching while reading. This is the "twist" mentioned
in the title of this PR. When one of these inferences happens, you'll see
a warning in the console (in development).

You do have to opt into fuzzy matching, but it can be a useful strategy
for preparing your clients for upcoming server changes, or for relaxing
the rules for a particular supertype, in cases when you know that its
subtypes change frequently on the server.